feat: add info tooltip to expected month-end balance
