+++ i18n/de-DE/finance.json
@@ -17,8 +17,9 @@
         "account": {
             "transactionHistory": "Transaktionsverlauf",
             "balance": "Kontostand",
-            "currentBalance": "Aktueller Kontostand",
-            "expectedBalance": "Erwarteter Kontostand zum Monatsende",
+            "currentBalance": "Aktuell",
+            "expectedBalance": "Erwartet",
+            "expectedBalanceInfo": "Berechnet aus allen kommenden geplanten Transaktionen und allen manuellen Transaktionen, die du letzten Monat nach dem aktuellen Kalendertag erstellt hast.",
             "addTransaction": "Transaktion hinzufügen",
             "manageAutoTransactions": "Automatische Transaktionen verwalten",
             "form": {
