+++ i18n/en-US/finance.json
@@ -17,8 +17,9 @@
         "account": {
             "transactionHistory": "Transaction History",
             "balance": "Account Balance",
-            "currentBalance": "Current account balance",
-            "expectedBalance": "Expected Month-End Balance",
+            "currentBalance": "Current",
+            "expectedBalance": "Expected",
+            "expectedBalanceInfo": "Calculated from all upcoming scheduled transactions and all manual transactions you made last month after the current day of the month.",
             "addTransaction": "Add transaction",
             "manageAutoTransactions": "Manage automatic transactions",
             "form": {
