+++ i18n/es-ES/finance.json
@@ -17,8 +17,9 @@
         "account": {
             "transactionHistory": "Historial de transacciones",
             "balance": "Saldo de la cuenta",
-            "currentBalance": "Saldo actual de la cuenta",
-            "expectedBalance": "Saldo previsto a fin de mes",
+            "currentBalance": "Actual",
+            "expectedBalance": "Previsto",
+            "expectedBalanceInfo": "Calculado a partir de todas las transacciones programadas futuras y de todas las transacciones manuales que hiciste el mes pasado después del día actual del mes.",
             "addTransaction": "Agregar transacción",
             "manageAutoTransactions": "Gestionar transacciones automáticas",
             "form": {
