+++ test-utils/playwright/snapshots/finance/account/balance-chart.test.ts-snapshots/balance-chart.aria.yml
@@ -1,1 +1,1 @@
-- text: "Current account balance: -€395.17 Expected Month-End Balance: -€395.17"
+- text: "Current: -€395.17 Expected: €187.06"
