+++ test-utils/playwright/snapshots/finance/other/filters.test.ts-snapshots/account-tab-filtered-by-category.aria.yml
@@ -12,7 +12,7 @@
       - button "Add transaction"
       - button "Manage automatic transactions"
       - button "Filters"
-      - text: "/Current account balance: €\\d+\\.\\d+ Expected Month-End Balance: €\\d+\\.\\d+/"
+      - text: "Current: €0.00 Expected: €901.85"
       - heading "Transaction History" [level=2]
       - list:
         - listitem:
